less code duplication + add enhanced determinism check on ci
